Minor cleanups suggested on my blog.

+++ src/core/commandqueue.cpp
@@ -64,7 +64,7 @@ cl_int CommandQueue::info(cl_context_info param_name,
                           size_t *param_value_size_ret)
 {
     void *value = 0;
-    int value_length = 0;
+    size_t value_length = 0;
     
     union {
         cl_uint cl_uint_var;
@@ -508,7 +508,7 @@ cl_int Event::info(cl_context_info param_name,
                    size_t *param_value_size_ret)
 {
     void *value = 0;
-    int value_length = 0;
+    size_t value_length = 0;
     
     union {
         cl_command_queue cl_command_queue_var;
