+++ src/core/cpudevice.cpp
@@ -180,7 +180,7 @@ cl_int CPUDevice::info(cl_device_info param_name,
                        size_t *param_value_size_ret)
 {
     void *value = 0;
-    int value_length = 0;
+    size_t value_length = 0;
     
     union {
         cl_device_type cl_device_type_var;
